Move current_version to sdr
This will allow us to have a current_version call that references the VersionMetadataDS
datastream
