+++ lib/dor/services/client/objects.rb
@@ -1,12 +1,15 @@
 # frozen_string_literal: true
 
 require 'nokogiri'
+require 'deprecation'
 
 module Dor
   module Services
     class Client
       # API calls that are about a repository object
       class Objects < VersionedService
+        extend Deprecation
+
         # Creates a new object in DOR
         # @return [HashWithIndifferentAccess] the response, which includes a :pid
         def register(params:)
@@ -46,16 +49,9 @@ module Dor
         # @raise [MalformedResponse] when the response is not parseable.
         # @return [Integer] the current version
         def current_version(object:)
-          xml = current_version_response(object: object)
-          begin
-            doc = Nokogiri::XML xml
-            raise if doc.root.name != 'currentVersion'
-
-            return Integer(doc.text)
-          rescue StandardError
-            raise MalformedResponse, "Unable to parse XML from current_version API call: #{xml}"
-          end
+          SDR.new(connection: connection, version: api_version).current_version(object: object)
         end
+        deprecation_deprecate current_version: 'Use Dor::Client.sdr.current_version instead'
 
         private
 
@@ -73,22 +69,6 @@ module Dor
           return resp.body if resp.success?
 
           raise UnexpectedResponse, "#{resp.reason_phrase}: #{resp.status} (#{resp.body})"
-        end
-
-        # make the request to the server for the currentVersion xml
-        # @raises [UnexpectedResponse] on an unsuccessful response from the server
-        # @returns [String] the raw xml from the server
-        def current_version_response(object:)
-          resp = connection.get do |req|
-            req.url current_version_path(object: object)
-          end
-          return resp.body if resp.success?
-
-          raise UnexpectedResponse, "#{resp.reason_phrase}: #{resp.status} (#{resp.body}) for #{object}"
-        end
-
-        def current_version_path(object:)
-          "#{api_version}/sdr/objects/#{object}/current_version"
         end
       end
     end
